+++ src/components/Footer/Footer.module.css
@@ -1,0 +1,5 @@
+.footer {
+  display: flex;
+  justify-content: flex-end;
+  padding: 10px;
+}
