+++ src/components/Header/Header.module.css
@@ -1,0 +1,5 @@
+.header {
+  display: flex;
+  justify-content: flex-end;
+  padding: 10px;
+}
