+++ src/components/Header/LanguageToggler/LanguageToggler.module.css
@@ -1,0 +1,2 @@
+.languageToggler {
+}
