--- src/components/LanguageToggler/LanguageToggler.module.css
@@ -1,5 +1,0 @@
-.wrapper {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-}
