+++ src/components/Main/Converter/Converter.module.css
@@ -2,4 +2,5 @@
   display: flex;
   flex-direction: column;
   gap: 50px;
+  align-items: center;
 }
