+++ src/components/Main/Converter/Output/values/HexValue.tsx
