+++ src/components/Main/Main.module.css
@@ -1,0 +1,7 @@
+.main {
+  flex-grow: 1;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
