+++ src/components/Main/index.tsx
@@ -1,0 +1,12 @@
+import { FC } from 'react';
+
+import Converter from './Converter';
+import styles from './Main.module.css';
+
+const Main: FC = () => (
+  <main className={styles.main}>
+    <Converter />
+  </main>
+);
+
+export default Main;
