+++ src/tests/components/Converter/Converter.test.tsx
@@ -1,11 +1,10 @@
-import Converter from '@components/Converter';
+import Converter from '@components/Main/Converter';
 import i18n from '@shared/i18n';
 import { fireEvent, render } from '@testing-library/react';
 import { I18nextProvider } from 'react-i18next';
 import { describe, expect, it } from 'vitest';
 
 import cardNumbers from '../../mock/cardNumbers.json';
-
 
 describe('Converter Component', () => {
   it('should render the Converter component', () => {
